Design messages refactor v2
