+++ src/app/pages/home/Employees/modals/ModalEmployees.js
@@ -216,6 +216,7 @@ const ModalEmployees = ({
             })
           ),
           postDB('messages', {
+            formatDate: formatDate,
             from: messageFrom,
             html: layout,
             read: read,
