--- src/app/pages/home/Messages/components/GeneralMessageContainer.scss
@@ -1,63 +1,0 @@
-.__container-gmc {
-  display: flex;
-  margin: 0;
-  min-height: 600px;
-  padding: 0;
-  width: 100%;
-  height: 620px;
-}
-
-.__container-general-snapshot {
-  height: 600px;
-  overflow: hidden;
-  width: 350px;
-}
-
-.snapshot-color {
-  background-color: rgba($color: #474747, $alpha: 0.2);
-}
-
-.snapshot-click {
-  background-color: purple;
-}
-
-.__container-general-snapshot::-webkit-scrollbar {
-  -webkit-appearance: none;
-}
-
-.__container-general-snapshot::-webkit-scrollbar:vertical {
-  width: 10px;
-}
-
-.__container-general-snapshot::-webkit-scrollbar-button:increment,
-.__container-general-snapshot::-webkit-scrollbar-button {
-  display: none;
-}
-
-.__container-general-snapshot::-webkit-scrollbar:horizontal {
-  height: 10px;
-}
-
-.__container-general-snapshot::-webkit-scrollbar-thumb {
-  border: 2px solid #f1f2f3;
-  background-color: #797979;
-  border-radius: 20px;
-}
-
-.__container-general-snapshot::-webkit-scrollbar-track {
-  border-radius: 10px;
-}
-
-.__container-snapshot {
-  margin: 15px;
-  min-height: 100px;
-}
-
-.__container-preview {
-  min-height: 600px;
-  width: 76%;
-}
-
-.__container-information {
-  min-height: 100px;
-}
