--- src/app/pages/home/Messages/components/MessageInformation.scss
@@ -1,8 +1,0 @@
-.message-information-img {
-    border-radius: 50%;
-    width: 80px;
-}
-
-.container-messages-preview {
-    padding-left: 24px;
-}
