--- src/app/pages/home/Messages/components/Snapshot.scss
@@ -1,32 +1,0 @@
-.container-messages-snapshot {
-  position: relative;
-  min-width: 270px;
-}
-
-.container-messages-snapshot:hover {
-  background-color: rgba($color: #302c2c8c, $alpha: 0.1);
-  cursor: pointer;
-}
-
-.container-messages-snapshot-selected {
-  background-color: rgba($color: #302c2c8c, $alpha: 0.1);
-  cursor: pointer;
-}
-
-.container-messages-snapshot:hover > .container-snapshot-delete-icon {
-  display: block;
-}
-
-.container-snapshot-delete-icon {
-  color: black;
-  display: none;
-  position: absolute;
-  right: 50px;
-  top: 30%;
-  width: 1px;
-}
-
-.snapshot-delete-icon:hover {
-  color: red;
-  cursor: pointer;
-}
