--- src/app/pages/home/Messages/components/TrashMessagesContainer.scss
@@ -1,77 +1,0 @@
-.trash-container {
-  display: flex;
-  flex-direction: row;
-  margin: 0;
-  height: 600px;
-}
-
-.messages-column-container {
-  display: flex;
-  flex-direction: column;
-  flex: 1;
-}
-
-.message-preview-container {
-  display: flex;
-  flex: 2;
-  align-items: center;
-  justify-content: center;
-}
-
-.search-box-container {
-  display: flex;
-  width: 90%;
-  height: 40px;
-  padding: 0px 10px;
-  align-self: flex-start;
-  align-items: center;
-  background-color: #fafafa;
-  border: 1px solid #E7E7E7;
-  border-radius: 10px;
-  &:hover {
-    background-color: #F5F5F5,;
-  }
-  svg {
-    margin-right: 10px;
-  }
-  input {
-    width: 100%;
-    border: none;
-    outline: none;
-    background-color: transparent;
-  }
-}
-
-.messages-container {
-  display: flex;
-  flex-direction: column;
-  flex: 11;
-  align-items: center;
-  justify-content: center;
-  margin-top: 10px;
-  overflow-y: scroll;
-}
-
-.snapshot-message {
-  display: flex;
-  align-self: center;
-  flex-direction: row;
-  align-items: center;
-  width: 90%;
-  padding-left: 20px;
-}
-
-.snapshot-message-info {
-  height: 100%;
-  align-content: space-around;
-  display: flex;
-  flex: 1;
-  flex-direction: column;
-}
-
-.page-controller-container {
-  display: flex;
-  flex: 1;
-  align-items: center;
-  justify-content: flex-start;
-}
